# src/routes/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Image Processing Swagger - Advanced Full-Stack Web Development Nanodegree Program",
        "version": "1.0.0",
        "title": "Image Processing Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": { "email": "hamed.farag.2009@gmail.com" },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/api/v1",
    "tags": ["udacity", "nanodegree", "Advanced", "full-stack"],
    "schemes": ["http"],
    "paths": {
        "/images": {
            "post": {
                "tags": ["image"],
                "summary": "uploads an image",
                "description": "Process your image by upload it using this API",
                "operationId": "postImage",
                "consumes": ["multipart/form-data"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "width",
                        "in": "formData",
                        "description": "Width of the new processed Image",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "height",
                        "in": "formData",
                        "description": "Height of the new processed Image",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "Image to upload",
                        "required": false,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    }
                }
            },
            "get": {
                "tags": ["image"],
                "summary": "get already uploaded image",
                "description": "Get thumbnail image from already uploaded image before",
                "operationId": "getImage",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "width",
                        "in": "query",
                        "description": "Width of the new processed Image",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "height",
                        "in": "query",
                        "description": "Height of the new processed Image",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "filename",
                        "in": "query",
                        "description": "Image to upload",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation"
                    }
                }
            }
        }
    },
    "externalDocs": { "description": "Find out more about Swagger", "url": "http://swagger.io" }
}